% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertqsUtils.R
\name{data2insertmssqlqs}
\alias{data2insertmssqlqs}
\title{dataframe 2 mssql insert query string}
\usage{
data2insertmssqlqs(da, tbn, header = FALSE)
}
\arguments{
\item{da, }{dataframe}

\item{tbn, }{table name}

\item{header, }{wheather include colnames or not}
}
\value{
mssql insert query string
}
\description{
dataframe 2 mssql insert query string
}
