% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertqsUtils.R
\name{data2insertmysqlqs}
\alias{data2insertmysqlqs}
\title{dataframe 2 mysql insert query string}
\usage{
data2insertmysqlqs(da, tbn)
}
\arguments{
\item{da, }{dataframe}

\item{tbn, }{table name}
}
\value{
mysql insert query string
}
\description{
dataframe 2 mysql insert query string
}
