% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetQueryForkExec.R
\name{dbGetQueryForkExec}
\alias{dbGetQueryForkExec}
\title{A function that uses the parallel package to execute a SQL statement in a separate process.}
\usage{
dbGetQueryForkExec(dbConnectionFun, dbConnectionArgs, statement,
  async = FALSE)
}
\arguments{
\item{dbConnectionFun}{The function to be called to instantiate a database connection}

\item{dbConnectionArgs}{The values of the parameters (in a named list) to be passed to dbConnectionFun to instantiate a connection}

\item{statement}{The SQL statement to be executed}

\item{async}{If FALSE, then the function blocks and waits for the subprocess to complete, and returns the result of the query.  If TRUE, 
the function asyncronously returns an object of the class parallelJob that can be used to collect the results via parallel::mccollect}
}
\description{
A function that uses the parallel package to execute a SQL statement in a separate process.
}
\keyword{sql}
