% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDbGetQueryForkExec.R
\name{doDbGetQuery}
\alias{doDbGetQuery}
\title{A helper function that is called by the fork'd child process to execute the SQL statement.}
\usage{
doDbGetQuery(params)
}
\description{
A helper function that is called by the fork'd child process to execute the SQL statement.
}
\keyword{sql}
