% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDbSendUpdateForkExec.R
\name{doDbSendUpdate}
\alias{doDbSendUpdate}
\title{A helper function that is called by the fork'd child process to execute the SQL statement.}
\usage{
doDbSendUpdate(params)
}
\description{
A helper function that is called by the fork'd child process to execute the SQL statement.
}
\keyword{sql}
