% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnnameUtils.R
\name{ms_columnnames}
\alias{ms_columnnames}
\title{to MSSQL column names}
\usage{
ms_columnnames(columnnames)
}
\arguments{
\item{columnnames, }{vector of columnnames}
}
\value{
a vector of MSSQL compatible columnnames
}
\description{
let dataframe columnnames fit the constraint of MSSQL column names
}
