% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemaUtils.R
\name{ms_qsg_schema}
\alias{ms_qsg_schema}
\title{use R dataframe or tibble to generate MSSQL table schema}
\usage{
ms_qsg_schema(dT, tablename, lnvarchar = 50)
}
\arguments{
\item{dT, }{R data.frame or tibble}

\item{tablename, }{MSSQL table name}

\item{lvarchar, }{length of varchar, default=50}
}
\description{
use R dataframe or tibble to generate MSSQL table schema
}
