% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnnameUtils.R
\name{my_columnnames}
\alias{my_columnnames}
\title{to MySQL column names}
\usage{
my_columnnames(columnnames)
}
\arguments{
\item{columnnames, }{vector of columnnames}
}
\value{
a vector of MySQL compatible columnnames
}
\description{
let dataframe columnnames fit the constraint of MySQL column names
}
