% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysqlutils.R
\name{my_qsg_collateutf8}
\alias{my_qsg_collateutf8}
\title{get queryString to set mysql table column collate as utf8_coration_ci}
\usage{
my_qsg_collateutf8(tablename, columnname, charlength = 50)
}
\arguments{
\item{tablename, }{mysql table name}

\item{columnname, }{table column name}

\item{charlength, }{varchar length default=50}
}
\value{
mysql alter table sql string
}
\description{
get queryString to set mysql table column collate as utf8_coration_ci
}
