% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemaUtils.R
\name{my_qsg_schema}
\alias{my_qsg_schema}
\title{use R dataframe or tibble to generate MySQL table schema}
\usage{
my_qsg_schema(dT, tablename, lnvarchar = 50)
}
\arguments{
\item{dT, }{R data.frame or tibble}

\item{tablename, }{MySQL table name}

\item{lvarchar, }{length of varchar, default=50}
}
\description{
use R dataframe or tibble to generate MySQL table schema
}
