% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnnameUtils.R
\name{pg_columnnames}
\alias{pg_columnnames}
\title{to PostgreSQL column names}
\usage{
pg_columnnames(columnnames)
}
\arguments{
\item{columnnames, }{vector of columnnames}
}
\value{
a vector of PostgreSQL compatible columnnames
}
\description{
let dataframe columnnames fit the constraint of PostgreSQL column names
}
