% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertqsUtils.R
\name{pg_qsg_fromdf}
\alias{pg_qsg_fromdf}
\title{dataframe or tibble to PostgreSQL insert query string}
\usage{
pg_qsg_fromdf(dT, tablename, iLength = 1024, tz = Sys.timezone(),
  Colnames = TRUE)
}
\arguments{
\item{dT, }{a tibble}

\item{tablename, }{table name}

\item{iLength, }{max size(row) counts in KB, default = 1024}

\item{tz, }{timezone}

\item{Colnames, }{boolean whether include colnames inside query string, default = TRUE}
}
\value{
a tibble of a segmented insert query string
}
\description{
dataframe or tibble to PostgreSQL insert query string
}
\examples{
\dontrun{
  require(tidyverse) 
  mytbl <- tibble(a=c(1, 2, 3, 4), b=c('1', '2', '3', '4'), 
    c=c(Sys.Date(), Sys.Date(), Sys.Date(), Sys.Date()))
  pg_qsg_fromdf(mytbl, 'tmp') 
} 
}
