% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemaUtils.R
\name{pg_qsg_schema}
\alias{pg_qsg_schema}
\title{use R dataframe or tibble to generate PostgreSQL table schema}
\usage{
pg_qsg_schema(dT, tablename, lnvarchar = 50)
}
\arguments{
\item{dT, }{R data.frame or tibble}

\item{tablename, }{PostgreSQL table name}

\item{lvarchar, }{length of varchar, default=50}
}
\description{
use R dataframe or tibble to generate PostgreSQL table schema
}
