% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phpmyadmin_webapi.R
\name{pma_getquery}
\alias{pma_getquery}
\title{get a query result of mysql}
\usage{
pma_getquery(queryString, dbname, Url, Token)
}
\arguments{
\item{queryString, }{query string}

\item{dbname, }{database name}

\item{Url, }{remote server URL}

\item{Token, }{login token}
}
\value{
tibble
}
\description{
get a query result of mysql
}
