% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phpmyadmin_webapi.R
\name{pma_gettable}
\alias{pma_gettable}
\title{show tables on phpmyadmin server}
\usage{
pma_gettable(dbname, Url, Token)
}
\arguments{
\item{dbname, }{database name}

\item{Url, }{remote server URL}

\item{Token, }{login token}
}
\value{
table names
}
\description{
show tables on phpmyadmin server
}
