% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phpmyadmin_webapi.R
\name{pma_gettoken}
\alias{pma_gettoken}
\title{get phpmyadmin login token}
\usage{
pma_gettoken(user, pass, Url)
}
\arguments{
\item{user, }{phpmyadmin username}

\item{pass, }{phpmyadmin user password}

\item{Url, }{remote server URL}
}
\value{
pma_token
}
\description{
get phpmyadmin login token
}
\examples{
\dontrun{
  mytoken <- pma_gettoken('test', '****', 'localhost:8080/phpmyadmin') 
  mytoken 
} 
}
