% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phpmyadmin_webapi.R
\name{pma_sendquery}
\alias{pma_sendquery}
\title{send a SQL query string to phpmyadmin}
\usage{
pma_sendquery(queryString, dbname, Url, Token)
}
\arguments{
\item{queryString, }{SQL query string. Be care of the string length and string length each line.}

\item{dbname, }{database name}

\item{Url, }{remote server URL}

\item{Token, }{login token}
}
\value{
http status code
}
\description{
send a SQL query string to phpmyadmin
}
\examples{
\dontrun{
  mytoken <- pma_gettoken('test', '****', 'localhost:8080/phpmyadmin')
  pma_sendquery('truncate table iris', 'testdb', 'localhost:8080/phpmyadmin', mytoken)
}
}
