% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqlstring}
\alias{sqlstring}
\title{sqlstring generic}
\usage{
sqlstring(x, ...)
}
\arguments{
\item{x}{object to sqlstring}

\item{tz}{timezone}
}
\value{
sqlstring
}
\description{
sqlstring generic
}
\examples{
  require(tidyverse) 
  mytbl <- tibble::tibble(c1 = c(1, 2, 3), c2 = c(1.1, 2.1, 3.1), c3 = c(1, NA_real_, 3),
    c4 = c('A', 'B', 'C') , c5 = c(Sys.time(), Sys.time(), Sys.time()), 
    c5 = c(Sys.Date(), NA, Sys.Date), c6 = c(NA, NA, NA))
 mytbl \%>\% mutate_all(funs(sqlstring))
 mytbl \%>\% mutate_all(funs(sqlstring), tz="US/Eastern") 
}
