% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_syncexportmssql.R
\name{syncmssql}
\alias{syncmssql}
\title{syncmssql}
\usage{
syncmssql(tablename, filename, qsfunc, dbconnection, ...)
}
\arguments{
\item{tablename, }{export tablename}

\item{filename, }{export data filename}

\item{qsfunc, }{querystring generate function}

\item{dbconnection, }{database connection}
}
\description{
syncmssql
}
