% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_syncexportmssql.R
\name{syncqsdaily}
\alias{syncqsdaily}
\title{syncqsdaily}
\usage{
syncqsdaily(tablename, columnname, ...)
}
\arguments{
\item{tablename, }{tablename}

\item{columnname, }{columnname}
}
\description{
syncqsdaily
}
