% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threaded.chunked.query.R
\name{threaded.chunked.query}
\alias{threaded.chunked.query}
\title{A function to pull a large data table from SQL Server to R in parallel chunks
This function relies on the parallel package which uses a fork / exec mechanism.
Therefore, it is important that the parent process not have an initialized JVM,
as the child processes will see it and attempt to use it.}
\usage{
threaded.chunked.query(database, user, password, server.address, domain,
  table.name, order.by.clause, n.threads = 8)
}
\arguments{
\item{database}{The name of the database to connect to}

\item{user}{The name of the user for the database}

\item{password}{The database user's password}

\item{server.address}{The network address (DNS or IP) of the database server}

\item{domain}{The Windows domain for the specified users}

\item{table.name}{The name of the table to be pulled from the database}

\item{order.by.clause}{The ORDER BY clause that will be used to determine the order in which
rows are fetched.  This MUST contain a column or combination of columns that are guaranteed have
unique values for every row, or the behavior of the FETCH operation may not be predictable.}

\item{n.threads}{The number of parallel jobs to be used to fetch the table}
}
\description{
A function to pull a large data table from SQL Server to R in parallel chunks
This function relies on the parallel package which uses a fork / exec mechanism.
Therefore, it is important that the parent process not have an initialized JVM,
as the child processes will see it and attempt to use it.
}
\keyword{sql}
