% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tibbleiqs}
\alias{tibbleiqs}
\title{tibble sqlstring cell}
\usage{
tibbleiqs(dT, tz = Sys.timezone())
}
\arguments{
\item{dT, }{a tibble}

\item{tz, }{timezone}
}
\value{
a list contains QueryStrings, Colnames, RowKB
}
\description{
tibble sqlstring cell
}
