% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetQueryForkExec.R
\name{dbGetQueryForkExec}
\alias{dbGetQueryForkExec}
\title{A function that uses the parallel package to execute a SQL statement in a separate process,
thereby avoiding creating a JVM in the current process.}
\usage{
dbGetQueryForkExec(database, user, password, server.address, domain, statement)
}
\arguments{
\item{database}{The name of the database to connect to}

\item{user}{The name of the user for the database}

\item{password}{The database user's password}

\item{server.address}{The network address (DNS or IP) of the database server}

\item{domain}{The Windows domain for the specified users}

\item{statement}{The SQL statement to be issued in a dbGetQuery(...) call}
}
\description{
A function that uses the parallel package to execute a SQL statement in a separate process,
thereby avoiding creating a JVM in the current process.
}
\keyword{sql}
